+++ app/javascript/src/js/previewShare.js
