+++ app/javascript/src/js/previewShare.test.js
@@ -1,14 +1,16 @@
-import { generateLink, encodeContent, decodeContent } from './mdPreviewShare';
+import { generateLink, encodeContent, decodeContent } from './previewShare';
 
 describe('encodeContent', () => {
   it('content is correctly encoded to base64', () => {
     const buffer = Buffer.from('This is a test string!', 'utf-8');
     expect(encodeContent('This is a test string!')).toBe(buffer.toString('base64'));
   });
+
   it('content including numbers is correctly encoded to base64', () => {
     const buffer = Buffer.from('This is a 2nd test string!', 'utf-8');
     expect(encodeContent('This is a 2nd test string!')).toBe(buffer.toString('base64'));
   });
+
   it('content including special characters and new lines correctly encoded to base64', () => {
     const buffer = Buffer.from('This is a test string! \n # Next Title', 'utf-8');
     expect(encodeContent('This is a test string! \n # Next Title')).toBe(buffer.toString('base64'));
@@ -20,10 +22,12 @@ describe('decodeContent', () => {
     const buffer = Buffer.from('This is a test string!', 'utf-8').toString('base64');
     expect(decodeContent(buffer)).toBe('This is a test string!');
   });
+
   it('content including numbers is correctly decoded from base64', () => {
     const buffer = Buffer.from('This is a 2nd test string!', 'utf-8').toString('base64');
     expect(decodeContent(buffer)).toBe('This is a 2nd test string!');
   });
+
   it('content including special characters and new lines correctly decoded from base64', () => {
     const buffer = Buffer.from('This is a test string! \n # Next Title', 'utf-8').toString('base64');
     expect(decodeContent(buffer)).toBe('This is a test string! \n # Next Title');
